requirements.txt
meross_iot==0.4.1.0rc7
apscheduler==3.6.3
telepot>=12.7
mysql-connector-python>=8.0
